Restructuring bot as subpackage
