+++ docs/bot.rst
@@ -1,8 +1,25 @@
-Bot Module
-==========
+Bot Subpackage
+==============
 
-This module contains the classes and functions used to run a YugiQuery bot.
+This subpackage contains the bot related modules used to run a YugiQuery bot.
 
-.. automodule:: yugiquery.bot
+Base Module
+-----------
+
+.. automodule:: yugiquery.bot.base
+   :members:
+   :show-inheritance:
+
+Discord Module
+--------------
+
+.. automodule:: yugiquery.bot.discord
+   :members:
+   :show-inheritance:
+
+Telegram Module
+---------------
+
+.. automodule:: yugiquery.bot.telegram
    :members:
    :show-inheritance:
